Last changes of todos cards
